Backend security + user profiles secured in the REST controller API, there are some more GUI improvement when displaying the restaurants(dishes to be inserted)

+++ src/main/java/com/gastronomee/web/rest/MenuResource.java
@@ -1,14 +1,14 @@
 package com.gastronomee.web.rest;
 
-import com.codahale.metrics.annotation.Timed;
-import com.gastronomee.domain.Menu;
+import static org.elasticsearch.index.query.QueryBuilders.queryStringQuery;
 
-import com.gastronomee.repository.MenuRepository;
-import com.gastronomee.repository.search.MenuSearchRepository;
-import com.gastronomee.web.rest.util.HeaderUtil;
-import com.gastronomee.web.rest.util.PaginationUtil;
-import io.swagger.annotations.ApiParam;
-import io.github.jhipster.web.util.ResponseUtil;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -16,17 +16,30 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
+import com.codahale.metrics.annotation.Timed;
+import com.gastronomee.domain.Menu;
+import com.gastronomee.domain.Restaurant;
+import com.gastronomee.repository.MenuRepository;
+import com.gastronomee.repository.RestaurantRepository;
+import com.gastronomee.repository.search.MenuSearchRepository;
+import com.gastronomee.security.AuthoritiesConstants;
+import com.gastronomee.security.SecurityUtils;
+import com.gastronomee.web.rest.util.HeaderUtil;
+import com.gastronomee.web.rest.util.PaginationUtil;
 
-import static org.elasticsearch.index.query.QueryBuilders.*;
+import io.github.jhipster.web.util.ResponseUtil;
+import io.swagger.annotations.ApiParam;
 
 /**
  * REST controller for managing Menu.
@@ -38,14 +51,17 @@ public class MenuResource {
     private final Logger log = LoggerFactory.getLogger(MenuResource.class);
 
     private static final String ENTITY_NAME = "menu";
+    
+    private final RestaurantRepository restaurantRepository;
         
     private final MenuRepository menuRepository;
 
     private final MenuSearchRepository menuSearchRepository;
 
-    public MenuResource(MenuRepository menuRepository, MenuSearchRepository menuSearchRepository) {
+    public MenuResource(MenuRepository menuRepository, MenuSearchRepository menuSearchRepository, RestaurantRepository restaurantRepository) {
         this.menuRepository = menuRepository;
         this.menuSearchRepository = menuSearchRepository;
+        this.restaurantRepository=restaurantRepository;
     }
 
     /**
@@ -57,6 +73,10 @@ public class MenuResource {
      */
     @PostMapping("/menus")
     @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER
+    })
     public ResponseEntity<Menu> createMenu(@Valid @RequestBody Menu menu) throws URISyntaxException {
         log.debug("REST request to save Menu : {}", menu);
         if (menu.getId() != null) {
@@ -80,6 +100,10 @@ public class MenuResource {
      */
     @PutMapping("/menus")
     @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER
+    })
     public ResponseEntity<Menu> updateMenu(@Valid @RequestBody Menu menu) throws URISyntaxException {
         log.debug("REST request to update Menu : {}", menu);
         if (menu.getId() == null) {
@@ -102,8 +126,38 @@ public class MenuResource {
     @Timed
     public ResponseEntity<List<Menu>> getAllMenus(@ApiParam Pageable pageable) {
         log.debug("REST request to get a page of Menus");
-        Page<Menu> page = menuRepository.findAll(pageable);
+        
+        Page<Menu> page;
+        if (SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+        	page = menuRepository.findAll(pageable);//admin sees all menus
+        } else {
+        	List<Restaurant> restaurants = restaurantRepository.findByManagerIsCurrentUser();
+        	page = menuRepository.findAllByRestaurantIn(restaurants, pageable);//managers do not see each other menus
+        }
+             
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/menus");
+        return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
+    }
+    
+    /**
+     * GET  /menus : get all my menus.
+     *
+     * @param pageable the pagination information
+     * @return the ResponseEntity with status 200 (OK) and the list of menus in body
+     */
+    @GetMapping("/menus/my")
+    @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER
+    })
+    public ResponseEntity<List<Menu>> getMyMenus(@ApiParam Pageable pageable) {
+        log.debug("REST request to get my Menus");
+             
+        List<Restaurant> restaurants = restaurantRepository.findByManagerIsCurrentUser();
+        Page<Menu> page = menuRepository.findAllByRestaurantIn(restaurants, pageable);//managers do not see each other menus    
+             
+        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/menus/my");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
 
@@ -129,11 +183,24 @@ public class MenuResource {
      */
     @DeleteMapping("/menus/{id}")
     @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER
+    })
     public ResponseEntity<Void> deleteMenu(@PathVariable Long id) {
         log.debug("REST request to delete Menu : {}", id);
-        menuRepository.delete(id);
-        menuSearchRepository.delete(id);
-        return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+
+        List<Restaurant> restaurants = restaurantRepository.findByManagerIsCurrentUser();
+        List<Menu> menu = menuRepository.findAllByRestaurantIn(restaurants);//managers do not see each other menus  
+        
+        if (menu.stream().filter(m -> m.getId().equals(id)).count() > 0 || SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+            menuRepository.delete(id);
+            menuSearchRepository.delete(id);
+            return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        } else {//not allowed
+            return ResponseEntity.status(403).headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        }
+        
     }
 
     /**
