+++ src/main/java/com/gastronomee/web/rest/RestaurantResource.java
@@ -1,14 +1,14 @@
 package com.gastronomee.web.rest;
 
-import com.codahale.metrics.annotation.Timed;
-import com.gastronomee.domain.Restaurant;
+import static org.elasticsearch.index.query.QueryBuilders.queryStringQuery;
 
-import com.gastronomee.repository.RestaurantRepository;
-import com.gastronomee.repository.search.RestaurantSearchRepository;
-import com.gastronomee.web.rest.util.HeaderUtil;
-import com.gastronomee.web.rest.util.PaginationUtil;
-import io.swagger.annotations.ApiParam;
-import io.github.jhipster.web.util.ResponseUtil;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
@@ -16,17 +16,34 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.validation.Valid;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
+import com.codahale.metrics.annotation.Timed;
+import com.gastronomee.domain.Location;
+import com.gastronomee.domain.Menu;
+import com.gastronomee.domain.Restaurant;
+import com.gastronomee.repository.LocationRepository;
+import com.gastronomee.repository.MenuRepository;
+import com.gastronomee.repository.RestaurantRepository;
+import com.gastronomee.repository.UserRepository;
+import com.gastronomee.repository.search.LocationSearchRepository;
+import com.gastronomee.repository.search.RestaurantSearchRepository;
+import com.gastronomee.security.AuthoritiesConstants;
+import com.gastronomee.security.SecurityUtils;
+import com.gastronomee.web.rest.util.HeaderUtil;
+import com.gastronomee.web.rest.util.PaginationUtil;
 
-import static org.elasticsearch.index.query.QueryBuilders.*;
+import io.github.jhipster.web.util.ResponseUtil;
+import io.swagger.annotations.ApiParam;
 
 /**
  * REST controller for managing Restaurant.
@@ -42,10 +59,27 @@ public class RestaurantResource {
     private final RestaurantRepository restaurantRepository;
 
     private final RestaurantSearchRepository restaurantSearchRepository;
+    
+    private final MenuRepository menuRepository;
+    
+    private final LocationRepository locationRepository;
 
-    public RestaurantResource(RestaurantRepository restaurantRepository, RestaurantSearchRepository restaurantSearchRepository) {
+    private final LocationSearchRepository locationSearchRepository;
+    
+    private final UserRepository userRepository;
+
+    public RestaurantResource(RestaurantRepository restaurantRepository, 
+    		RestaurantSearchRepository restaurantSearchRepository,
+    		LocationRepository locationRepository, 
+    		LocationSearchRepository locationSearchRepository,
+    		MenuRepository menuRepository,
+    		UserRepository userRepository) {
         this.restaurantRepository = restaurantRepository;
         this.restaurantSearchRepository = restaurantSearchRepository;
+        this.locationRepository = locationRepository;
+        this.locationSearchRepository = locationSearchRepository;
+        this.menuRepository = menuRepository;
+        this.userRepository = userRepository;
     }
 
     /**
@@ -57,13 +91,29 @@ public class RestaurantResource {
      */
     @PostMapping("/restaurants")
     @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER,
+    })
     public ResponseEntity<Restaurant> createRestaurant(@Valid @RequestBody Restaurant restaurant) throws URISyntaxException {
         log.debug("REST request to save Restaurant : {}", restaurant);
         if (restaurant.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "idexists", "A new restaurant cannot already have an ID")).body(null);
         }
+        
+        //when a restaurant is created then a new location is created 
+        if(restaurant.getLocation()!=null){
+        	Location location = locationRepository.save(restaurant.getLocation());
+        	locationSearchRepository.save(location);
+        }
+        
+        
+        //the manager is the current user that is creating this restaurant
+        restaurant.setManager(userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin()).get());
+        
         Restaurant result = restaurantRepository.save(restaurant);
         restaurantSearchRepository.save(result);
+        
         return ResponseEntity.created(new URI("/api/restaurants/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert(ENTITY_NAME, result.getId().toString()))
             .body(result);
@@ -80,16 +130,42 @@ public class RestaurantResource {
      */
     @PutMapping("/restaurants")
     @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER,
+    })
     public ResponseEntity<Restaurant> updateRestaurant(@Valid @RequestBody Restaurant restaurant) throws URISyntaxException {
         log.debug("REST request to update Restaurant : {}", restaurant);
         if (restaurant.getId() == null) {
             return createRestaurant(restaurant);
         }
+        
+        if(restaurant.getLocation()!=null){
+        	Location location = locationRepository.save(restaurant.getLocation());
+        	locationSearchRepository.save(location);
+        }
+        
         Restaurant result = restaurantRepository.save(restaurant);
         restaurantSearchRepository.save(result);
+        
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, restaurant.getId().toString()))
             .body(result);
+    }
+    
+    /**
+     * GET  /restaurants/my : get all my restaurants.
+     *
+     * @param pageable the pagination information
+     * @return the ResponseEntity with status 200 (OK) and the list of restaurants in body
+     */
+    @GetMapping("/restaurants/my")
+    @Timed
+    public ResponseEntity<List<Restaurant>> getMyRestaurants(@ApiParam Pageable pageable) {
+        log.debug("REST request to get my Restaurants");
+        Page<Restaurant> page = restaurantRepository.findAllByManagerLogin(SecurityUtils.getCurrentUserLogin(), pageable);
+        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/restaurants/my");
+        return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
 
     /**
@@ -102,8 +178,30 @@ public class RestaurantResource {
     @Timed
     public ResponseEntity<List<Restaurant>> getAllRestaurants(@ApiParam Pageable pageable) {
         log.debug("REST request to get a page of Restaurants");
-        Page<Restaurant> page = restaurantRepository.findAll(pageable);
+       
+        Page<Restaurant> page = restaurantRepository.findAll(pageable);//admin sees all restaurants
+        /*if (SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+        	page = restaurantRepository.findAll(pageable);//admin sees all restaurants
+        } else if(SecurityUtils.) {
+        	page = restaurantRepository.findAllByManagerLogin(SecurityUtils.getCurrentUserLogin(), pageable);
+        }*/
+        
         HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/restaurants");
+        return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
+    }
+    
+    /**
+     * GET  /restaurant/:id/menus : get the menus of the restaurant with "id" .
+     *
+     * @param pageable the pagination information
+     * @return the ResponseEntity with status 200 (OK) and the list of menus in body
+     */
+    @GetMapping("/restaurants/{id}/menus")
+    @Timed
+    public ResponseEntity<List<Menu>> getAllMenusByRestaurant(@PathVariable Long id, @ApiParam Pageable pageable) {
+        log.debug("REST request to get a page of Menus");
+        Page<Menu> page = menuRepository.findAllByRestaurantId(id, pageable);
+        HttpHeaders headers = PaginationUtil.generatePaginationHttpHeaders(page, "/api/restaurants/{id}/menus");
         return new ResponseEntity<>(page.getContent(), headers, HttpStatus.OK);
     }
 
@@ -129,12 +227,24 @@ public class RestaurantResource {
      */
     @DeleteMapping("/restaurants/{id}")
     @Timed
+    @Secured({
+    	AuthoritiesConstants.ADMIN,
+    	AuthoritiesConstants.MANAGER,
+    })
     public ResponseEntity<Void> deleteRestaurant(@PathVariable Long id) {
         log.debug("REST request to delete Restaurant : {}", id);
-        restaurantRepository.delete(id);
-        restaurantSearchRepository.delete(id);
-        return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        
+        if (restaurantRepository.findOneByManagerLoginAndId(SecurityUtils.getCurrentUserLogin(), id) !=null || SecurityUtils.isCurrentUserInRole(AuthoritiesConstants.ADMIN)) {
+            restaurantRepository.delete(id);
+            restaurantSearchRepository.delete(id);
+            return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        } else {//not allowed
+            return ResponseEntity.status(403).headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
+        }
+        
     }
+    
+    
 
     /**
      * SEARCH  /_search/restaurants?query=:query : search for the restaurant corresponding
